Further centralization of data sheet
