+++ frontend/packages/sheet/src/module.styl
@@ -23,3 +23,19 @@ thead
   background: #999
   cursor: ew-resize
   z-index: 50
+
+.virtualized-sheet
+  position: absolute;
+  top: 0px
+  bottom: 0px
+  left 0em
+  right 0em
+  overflow-y scroll
+  overflow-x: hidden
+
+  .ui
+    position: sticky
+    top: 0px
+    bottom: 0px
+    left 0px
+    right 0px
