+++ frontend/src/admin/data-sheet/module.styl
@@ -25,21 +25,6 @@
     top 0
     background-color: var(--datasheet-read-only-background)
 
-.virtualized-sheet
-  position: absolute;
-  top: 0px
-  bottom: 0px
-  left 0em
-  right 0em
-  overflow-y scroll
-  overflow-x: hidden
-
-  .ui
-    position: sticky
-    top: 0px
-    bottom: 0px
-    left 0px
-    right 0px
 
 :global
   .sheet-header
